requirements.txt
pyahocorasick==1.4.0
Flask==1.1.1
flask-restful==0.3.7
